Fixing all exporting issues and test-running
