+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .c9/
 .idea/
 dist/
+lib/
