+++ .travis.yml
@@ -5,7 +5,7 @@ matrix:
     - node_js: "6"
 
 before_script:
-  - npm install
+  - npm run build-ts
 
 script:
   - npm test
