+++ src/aqtToQuery.ts
@@ -1,14 +1,14 @@
-import _ from 'lodash';
+import * as _ from 'lodash';
 
 /**
  * @example
- *   module.exports('name') // => 'name'
- *   module.exports(['name', 'surname', 'age']) // => 'name surname age '
- *   module.exports({ people: 'name', countries: ['flag']}) // => 'q0_42: people { name }q0_43: countries { flag  }'
- *   module.exports(['id', 'name', { coordinates: ['lat', 'long'] }, { test: ['a']}])
+ *   exports.default('name') // => 'name'
+ *   exports.default(['name', 'surname', 'age']) // => 'name surname age '
+ *   exports.default({ people: 'name', countries: ['flag']}) // => 'q0_42: people { name }q0_43: countries { flag  }'
+ *   exports.default(['id', 'name', { coordinates: ['lat', 'long'] }, { test: ['a']}])
  *   // => 'id name q2_42: coordinates { lat long  } q3_42: test { a  } '
  */
-module.exports = function queryTreeToGraphQLString(tree, parentIndex = 0) {
+export default function queryTreeToGraphQLString(tree, parentIndex = 0) {
   let output : string = '';
 
   if (_.isObject(tree) && !_.isArray(tree)) {
