+++ src/coverageCalculator.ts
@@ -1,13 +1,13 @@
 import schemaToQueryTree  from './schemaToQueryTree';
-import _ from 'lodash';
+import * as _ from 'lodash';
 
 const { getQueryFields, getQueryFieldsModes } = schemaToQueryTree;
 
 /**
  * @example
- *   module.exports('ObjectContainingTwoDeeplyNestedObjects', require('../test/unit/mockData'))
+ *   exports.default('ObjectContainingTwoDeeplyNestedObjects', require('../test/unit/mockData'))
  *   // => { coverageRatio: 1, notCoveredFields: []}
- *   module.exports('DeeplyNestedObjectWithPartialNoFollow', require('../test/unit/mockData'))
+ *   exports.default('DeeplyNestedObjectWithPartialNoFollow', require('../test/unit/mockData'))
  *   // => { coverageRatio: 0.5, notCoveredFields: ["DeeplyNestedObject___NOFollowPart", "DeeplyNestedObject___DeeplyNestedObject___DeepNest", "DeeplyNestedObject___ObjectField___NotSoDeepNest"]}
  */
 export default function coverageCalculator (rootName, schema) {
