+++ src/introspectionQuery.ts
