+++ src/queryGenerator.ts
