+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-      "outDir": "./dist/",        // path to output directory
+      "outDir": "./lib/",        // path to output directory
       "sourceMap": true,          // allow sourcemap support
       "strictNullChecks": true,   // enable strict null checks as a best practice
-      "module": "commonjs",            // specify module code generation
+      "module": "none",            // specify module code generation
       "moduleResolution": "node",
       "baseUrl": ".",
       "target": "es6",            // specify ECMAScript target version
@@ -16,6 +16,6 @@
       "allowJs": true             // allow a partial TypeScript and JavaScript codebase  
   },
   "include": [
-      "./lib/"
+      "./src/"
   ]
 }
